+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Enhancements
 - **Sheet Changes**
   - Sheet completion percentage is now based on visible responses only
+  - CSV exports now additionally split checkbox variables into individual columns by domain options
 - **Gem Changes**
   - Updated to rails 4.1.4
   - Updated to kaminari 0.16.1